Add files via upload
This program will have two features. The first feature will calculate the cost of skates for a customer at the CPCC Skate Shop. Total pricing will include the type of skate, the number of skates, and a 7.25% sales tax.

The second feature will randomly generate the amount of roller skates or roller blades being purchased and add a discount to the purchase based on the day of the week.